Moving application middleware setup out of app.js to improve readability
